Add PathData classes
Add a sealed PathData class and two subclasses:
  - AtomicPath
  - SegmentedPath

These two classes will be used to carry path data into the BaseRoute in a future commit.

This is part of an effort to improve that path processing within BaseRoute instead of relying on users to properly enter root and non-root paths.