Convert BaseRoute to use PathData
Convert the BaseRoute class to use PathData instead of relying on the
`path` property override.

This changes the routes to require a path in the constructor and gives
us the ability to "sanitize" and format the paths.

Routes automatically detect it they are root-level or a child of another
route and apply the leading slash accordingly.

When constructing routes, we can now use the default, `root`, or `join`
constructors to create routes without supplying any slashes ourselves.